update(front) Alluserlist, block list and friends list done, error handle wip

+++ frontend/src/social/userSocial.ts
@@ -1,11 +1,20 @@
 import { fetchApi } from "../api/fetch";
+import { fetchToken } from "../api/fetchToken";
 import { getUserInfo } from "../api/getterUser(s)";
 import { API_USER } from "../api/routes";
 // import { API_PEOPLE } from "../api/routes";
 import { alertTemporary } from "../components/ui/alert/alertTemporary";
+import { allUsersList } from "../pages/Profile/allUsersList";
+import { friendsList } from "../pages/Profile/friendsList";
+import { blockList } from "../pages/Profile/blockList";
+import { renderErrorPage } from "../controllers/renderPage";
 
 export async function handleUnfriend(target: HTMLElement) {
 	const user = await getUserInfo();
+	if (!user || user.status === "error") {
+		renderErrorPage('401');
+		return;
+	}
 
 	const response = await fetchApi(API_USER.SOCIAL.FRIENDS + `/${target.dataset.id}`, {
 		method: "DELETE",
@@ -15,16 +24,18 @@ export async function handleUnfriend(target: HTMLElement) {
 		return alertTemporary("error", "issues-with-friend-removal", user.data!.preferences!.theme, true);
 	}
 	alertTemporary("success", "friend-removed", user.data!.preferences!.theme, true);
-	window.location.reload();
+	target.parentElement?.parentElement?.parentElement?.remove();
+	document.getElementById("all-users-div")!.innerHTML = `${await allUsersList()}`;
 }
 
 export async function handleFriendRequest(target: HTMLElement, action: "send" | "accept") {
 
 	const user = await getUserInfo();
 	if (!user || user.status === "error") {
-		return window.location.href = "/login";
+		renderErrorPage('401');
+		return;
 	}
-	
+
 	const targetId = target.dataset.id;
 	const response = await fetchApi(API_USER.SOCIAL.PENDING + `${(action == "send" ? "" : "/accept")}/${targetId}`, {
 		method: "POST",
@@ -34,11 +45,12 @@ export async function handleFriendRequest(target: HTMLElement, action: "send" | 
 	});
 	if (response.status === "error") {
 
-		(action === "send") ? alertTemporary("error", "issues-with-friend-invitation", user.data!.preferences.theme, true) : alertTemporary("error", "issues-with-friend-acceptance", user.data!.preferences.theme, true);
+		(action === "send") ? alertTemporary("error", "issues-with-friend-invitation", user.data!.preferences!.theme, true) : alertTemporary("error", "issues-with-friend-acceptance", user.data!.preferences!.theme, true);
 
 		return;
 	}
-	(action === "send") ? alertTemporary("success", "friend-invitation-sent", user.data!.preferences.theme, true) : alertTemporary("success", "friend-invitation-accepted", user.data!.preferences.theme, true);
+	(action === "send") ? alertTemporary("success", "friend-invitation-sent", user.data!.preferences!.theme, true) : alertTemporary("success", "friend-invitation-accepted", user.data!.preferences!.theme, true);
+	target.parentElement?.remove();
 }
 
 export async function sendRefuseInvitation(target: HTMLElement) {
@@ -58,7 +70,8 @@ export async function sendRefuseInvitation(target: HTMLElement) {
 	}
 	alertTemporary("success", "friend-invitation-refused", user.data!.preferences!.theme, true);
 
-	target.remove();
+	target.parentElement?.remove();
+	document.getElementById("all-users-div")!.innerHTML = `${await allUsersList()}`;
 }
 
 export async function cancelFriendInvitation(target: HTMLElement) {
@@ -74,23 +87,33 @@ export async function cancelFriendInvitation(target: HTMLElement) {
 	target.parentElement?.remove();
 }
 
-export async function blockUser(target: HTMLElement) {
-	console.log("blockUser target", target);
+export async function blockUser(target: HTMLElement, isBlocking: boolean) {
+	
 	const user = await getUserInfo();
 	if (!user || user.status === "error") {
 		window.location.href = "/login";
 		return;
 	}
+	
 	const blockId = target.dataset.id;
-
 	const response = await fetchApi(API_USER.SOCIAL.BLOCKED + `/${blockId}`, {
-		method: "POST",
+		method: (isBlocking ? "DELETE" : "POST"),
 		body: JSON.stringify({})
 	});
+	
 	if (response.status === "error") {
-		alertTemporary("error", "issues-with-user-blocked", user.data!.preferences!.theme, true);
-		return;
+		return alertTemporary("error", "issues-with-user-blocked", user.data!.preferences!.theme, true);
 	}
-	alertTemporary("success", "user-blocked", user.data!.preferences!.theme, true);
-	window.location.reload();
+
+	alertTemporary("success", (isBlocking ? "user-unblocked" : "user-blocked"), user.data!.preferences!.theme, true);
+
+	const allUsersDiv = document.getElementById("all-users-div");
+	if (allUsersDiv) {
+		allUsersDiv.innerHTML = `${await allUsersList()}`;
+	}
+	const blockListDiv = document.getElementById("block-div");
+	if (blockListDiv) {
+		blockListDiv.innerHTML = `${await blockList()}`;
+	}
+
 }
